feat: Support delayed session assignment for builtin objects
